Mealplan calender and date implementation

+++ app/src/main/java/com/example/foodcart/mealplans/CustomMealPlanArrayAdapter.java
@@ -15,10 +15,15 @@ import androidx.annotation.Nullable;
 
 import com.example.foodcart.R;
 import com.example.foodcart.ingredients.Ingredient;
+import com.example.foodcart.ingredients.IngredientFragment;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
 
 import java.util.ArrayList;
 
@@ -58,9 +63,11 @@ public class CustomMealPlanArrayAdapter extends ArrayAdapter<Meal> {
         // Get the views of the TextViews
         TextView mealname = view.findViewById(R.id.mealplan_name);
         TextView mealtype = view.findViewById(R.id.mealplan_type);;
+        TextView mealdate = view.findViewById(R.id.mealplan_date);;
 
         mealname.setText(meal.getMealName());
         mealtype.setText(meal.getMealType());
+        mealdate.setText(meal.getFormattedDate());
 
         // set up delete button on each list item and onClick
         ImageButton deleteButton = (ImageButton) view.findViewById(R.id.mealplan_item_deleteButton);
@@ -78,23 +85,49 @@ public class CustomMealPlanArrayAdapter extends ArrayAdapter<Meal> {
                     db = FirebaseFirestore.getInstance();
                     // Get a top level reference to the collection
                     final CollectionReference MealPlanCollection = db.collection("MealPlan");
-                    MealPlanCollection
-                            .document(meals.get(position).getMealName())
-                            .delete()
-                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                                @Override
-                                public void onSuccess(Void aVoid) {
-                                    // These are a method which gets executed when the task is succeeded
-                                    Log.d("Delete MealPlan", "Data has been deleted successfully!");
-                                }
-                            })
-                            .addOnFailureListener(new OnFailureListener() {
-                                @Override
-                                public void onFailure(@NonNull Exception e) {
-                                    // These are a method which gets executed if there’s any problem
-                                    Log.d("Delete MealPlan", "Data could not be deleted!" + e.toString());
-                                }
-                            });
+                    Meal meal = meals.get(position);
+                    if(meal.getMealName().equals("Ingredient")) {
+                        MealPlanFragment.delMealIngredientDB(meal.getMealName(), MealPlanCollection);
+                    } else {
+                        // delete ingredient list
+                        MealPlanCollection
+                                .document(meal.getMealName())
+                                .collection("Ingredients")
+                                .get()
+                                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                    @Override
+                                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                        if (task.isSuccessful()) {
+                                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                                Log.d("Delete MealIngredient", document.getId() + meal.getMealName());
+                                                IngredientFragment.delIngredientDB(document.getId(),
+                                                        MealPlanCollection.document(meal.getMealName())
+                                                        .collection("Ingredients"));
+                                            }
+                                        } else {
+                                            Log.d("Delete Meal Ingredient", "Error getting documents: ", task.getException());
+                                        }
+                                    }
+                                });
+                        // delete recipe
+                        MealPlanCollection
+                                .document(meals.get(position).getMealName())
+                                .delete()
+                                .addOnSuccessListener(new OnSuccessListener<Void>() {
+                                    @Override
+                                    public void onSuccess(Void aVoid) {
+                                        // These are a method which gets executed when the task is succeeded
+                                        Log.d("Delete Meal Recipe", "Data has been deleted successfully!");
+                                    }
+                                })
+                                .addOnFailureListener(new OnFailureListener() {
+                                    @Override
+                                    public void onFailure(@NonNull Exception e) {
+                                        // These are a method which gets executed if there’s any problem
+                                        Log.d("Delete Meal Recipe", "Data could not be deleted!" + e.toString());
+                                    }
+                                });
+                    }
                     // find and remove selection
                     meals.remove(Math.min(position, meals.size() - 1));
                     notifyDataSetChanged();
